refactor: Remove DatabudgieManifestMixin in favor of a function to create the table definition.
