+++ src/databudgie/manifest/__init__.py
@@ -1,2 +1,2 @@
 # flake8: noqa
-from databudgie.manifest.mixins import DatabudgieManifestMixin
+from databudgie.manifest.sqlalchemy import create_manifest_table, DatabudgieManifestMixin
